ms_celery_ner/requirements.txt
celery
redis
eventlet
python-dotenv
nltk==3.7
tensorflow==2.10.0
torch==1.12.1
tokenizers==0.10.3
transformers==4.13.0
